requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
discord.py==1.5.1
google-auth==1.23.0
google-auth-oauthlib==0.4.2
gspread==3.6.0
idna==2.10
isort==5.6.4
joblib==0.17.0
keyboard==0.13.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
MouseInfo==0.1.3
multidict==4.7.6
numpy==1.19.3
oauthlib==3.1.0
Pillow==8.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
PyDirectInput==1.0.2
PyGetWindow==0.0.9
pylint==2.6.0
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
yarl==1.5.1
